scripts/requirements.txt
elevenlabs
python-dotenv

